+++ tests/Params/Tests/OptionsTraitTester.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Params\Tests;
+
+class OptionsTraitTester
+{
+    use \Params\OptionsTrait;
+}